+++ .gitignore
@@ -1,7 +1,7 @@
 /tmp
 /out-tsc
 
-/node_modules
+node_modules/
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
