+++ package.json
@@ -2,35 +2,8 @@
   "name": "5055",
   "version": "1.0.0",
   "description": "",
-  "main": "dist/index.js",
+  "private": true,
   "scripts": {
-    "build": "tsc",
-    "lint": "eslint . --max-warnings 0"
-  },
-  "dependencies": {
-    "@opentelemetry/exporter-logs-otlp-http": "^0.53.0",
-    "@opentelemetry/otlp-exporter-base": "^0.53.0",
-    "@opentelemetry/sdk-logs": "^0.53.0",
-    "tslib": "2.7.0"
-  },
-  "devDependencies": {
-    "eslint": "9.12.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-eslint-comments": "3.2.0",
-    "eslint-plugin-eslint-plugin": "6.2.0",
-    "eslint-plugin-import": "2.31.0",
-    "eslint-plugin-jest": "28.8.3",
-    "eslint-plugin-only-warn": "1.1.0",
-    "eslint-plugin-prettier": "5.2.1",
-    "eslint-plugin-simple-import-sort": "12.1.1",
-    "eslint-plugin-unicorn": "56.0.0",
-    "globals": "15.10.0",
-    "prettier": "3.3.3",
-    "rimraf": "6.0.1",
-    "ts-node": "10.9.2",
-    "tsup": "8.3.0",
-    "typescript": "5.6.2",
-    "typescript-eslint": "8.8.1"
-  },
-  "private": true
+  "lint": "pnpm --stream -r lint"
+}
 }
