+++ packages/core/eslint.config.mjs
@@ -1,0 +1,284 @@
+// @ts-check
+
+// Based on: https://github.com/typescript-eslint/typescript-eslint/blob/main/eslint.config.mjs
+
+import url from 'node:url';
+
+import eslint from '@eslint/js';
+import eslintCommentsPlugin from 'eslint-plugin-eslint-comments';
+import eslintPluginPlugin from 'eslint-plugin-eslint-plugin';
+import importPlugin from 'eslint-plugin-import';
+import jestPlugin from 'eslint-plugin-jest';
+import simpleImportSortPlugin from 'eslint-plugin-simple-import-sort';
+import unicornPlugin from 'eslint-plugin-unicorn';
+import globals from 'globals';
+import tseslint from 'typescript-eslint';
+
+import eslintPluginPrettierRecommended from 'eslint-plugin-prettier/recommended';
+
+const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
+
+export default tseslint.config(
+  // register all the plugins up-front
+  {
+    // note - intentionally uses computed syntax to make it easy to sort the keys
+    plugins: {
+      ['@typescript-eslint']: tseslint.plugin,
+      ['eslint-comments']: eslintCommentsPlugin,
+      ['eslint-plugin']: eslintPluginPlugin,
+      ['import']: importPlugin,
+      ['jest']: jestPlugin,
+      ['simple-import-sort']: simpleImportSortPlugin,
+      ['unicorn']: unicornPlugin,
+    },
+  },
+
+  // config with just ignores is the replacement for `.eslintignore`
+  {
+    ignores: [
+      '**/__snapshots__/**',
+      '**/coverage/**',
+      '**/build/**',
+      '**/dist/**',
+      '**/lib/**',
+      '**/fixtures/**',
+      '**/node_modules/**',
+      '**/*.mjs',
+      '**/*.mts',
+    ],
+  },
+
+  // prettier
+  eslintPluginPrettierRecommended,
+
+  // extends ...
+  eslint.configs.recommended,
+  ...tseslint.configs.strictTypeChecked,
+  ...tseslint.configs.stylisticTypeChecked,
+
+  // base config
+  {
+    ignores: ['**/test/integration/**/*'],
+    languageOptions: {
+      globals: {
+        ...globals.es2020,
+        ...globals.node,
+      },
+      parserOptions: {
+        allowAutomaticSingleRunInference: true,
+        project: ['tsconfig.lint.json', 'test/**/tsconfig*.json'],
+        tsconfigRootDir: __dirname,
+        warnOnUnsupportedTypeScriptVersion: false,
+      },
+    },
+    rules: {
+      // make sure we're not leveraging any deprecated APIs
+      '@typescript-eslint/no-deprecated': 'error',
+
+      // TODO(#7130): Investigate changing these in or removing these from presets
+      '@typescript-eslint/no-confusing-void-expression': 'off',
+      '@typescript-eslint/prefer-string-starts-ends-with': 'off',
+
+      '@typescript-eslint/ban-ts-comment': [
+        'error',
+        {
+          'ts-expect-error': 'allow-with-description',
+          'ts-ignore': true,
+          'ts-nocheck': true,
+          'ts-check': false,
+          minimumDescriptionLength: 5,
+        },
+      ],
+      '@typescript-eslint/consistent-type-imports': [
+        'error',
+        { prefer: 'type-imports', disallowTypeAnnotations: true },
+      ],
+      '@typescript-eslint/explicit-function-return-type': ['error', { allowIIFEs: true }],
+      '@typescript-eslint/no-explicit-any': 'error',
+      'no-constant-condition': 'off',
+      '@typescript-eslint/no-unnecessary-condition': ['error', { allowConstantLoopConditions: true }],
+      '@typescript-eslint/no-var-requires': 'off',
+      '@typescript-eslint/prefer-literal-enum-member': [
+        'error',
+        {
+          allowBitwiseExpressions: true,
+        },
+      ],
+      '@typescript-eslint/unbound-method': 'off',
+      '@typescript-eslint/restrict-template-expressions': [
+        'error',
+        {
+          allowNumber: true,
+          allowBoolean: true,
+          allowAny: true,
+          allowNullish: true,
+          allowRegExp: true,
+        },
+      ],
+      '@typescript-eslint/no-unused-vars': [
+        'error',
+        {
+          caughtErrors: 'all',
+          varsIgnorePattern: '^_',
+          argsIgnorePattern: '^_',
+        },
+      ],
+      '@typescript-eslint/prefer-nullish-coalescing': [
+        'error',
+        {
+          ignoreConditionalTests: true,
+          ignorePrimitives: true,
+        },
+      ],
+
+      //
+      // eslint-base
+      //
+
+      curly: ['error', 'all'],
+      eqeqeq: [
+        'error',
+        'always',
+        {
+          null: 'never',
+        },
+      ],
+      'logical-assignment-operators': 'error',
+      'no-else-return': 'error',
+      'no-mixed-operators': 'error',
+      'no-console': 'error',
+      'no-process-exit': 'error',
+      'no-fallthrough': ['error', { commentPattern: '.*intentional fallthrough.*' }],
+      'one-var': ['error', 'never'],
+
+      //
+      // eslint-plugin-eslint-comment
+      //
+
+      // require a eslint-enable comment for every eslint-disable comment
+      'eslint-comments/disable-enable-pair': [
+        'error',
+        {
+          allowWholeFile: true,
+        },
+      ],
+      // disallow a eslint-enable comment for multiple eslint-disable comments
+      'eslint-comments/no-aggregating-enable': 'error',
+      // disallow duplicate eslint-disable comments
+      'eslint-comments/no-duplicate-disable': 'error',
+      // disallow eslint-disable comments without rule names
+      'eslint-comments/no-unlimited-disable': 'error',
+      // disallow unused eslint-disable comments
+      'eslint-comments/no-unused-disable': 'error',
+      // disallow unused eslint-enable comments
+      'eslint-comments/no-unused-enable': 'error',
+      // disallow ESLint directive-comments
+      'eslint-comments/no-use': [
+        'error',
+        {
+          allow: ['eslint-disable', 'eslint-disable-line', 'eslint-disable-next-line', 'eslint-enable', 'global'],
+        },
+      ],
+
+      //
+      // eslint-plugin-import
+      //
+      // enforces consistent type specifier style for named imports
+      'import/consistent-type-specifier-style': 'error',
+      // disallow non-import statements appearing before import statements
+      'import/first': 'error',
+      // Require a newline after the last import/require in a group
+      'import/newline-after-import': 'error',
+      // Forbid import of modules using absolute paths
+      'import/no-absolute-path': 'error',
+      // disallow AMD require/define
+      'import/no-amd': 'error',
+      // forbid default exports - we want to standardize on named exports so that imported names are consistent
+      'import/no-default-export': 'error',
+      // disallow imports from duplicate paths
+      'import/no-duplicates': 'error',
+      // Forbid the use of extraneous packages
+      'import/no-extraneous-dependencies': [
+        'error',
+        {
+          devDependencies: true,
+          peerDependencies: true,
+          optionalDependencies: false,
+        },
+      ],
+      // Forbid mutable exports
+      'import/no-mutable-exports': 'error',
+      // Prevent importing the default as if it were named
+      'import/no-named-default': 'error',
+      // Prohibit named exports
+      'import/no-named-export': 'off', // we want everything to be a named export
+      // Forbid a module from importing itself
+      'import/no-self-import': 'error',
+      // Require modules with a single export to use a default export
+      'import/prefer-default-export': 'off', // we want everything to be named
+
+      // enforce a sort order across the codebase
+      'simple-import-sort/imports': 'error',
+    },
+  },
+
+  // Some rules must be turned off for JS sources
+  {
+    files: ['**/*.js'],
+
+    extends: [tseslint.configs.disableTypeChecked],
+    rules: {
+      // turn off other type-aware rules
+      'deprecation/deprecation': 'off',
+      '@typescript-eslint/internal/no-poorly-typed-ts-props': 'off',
+
+      // turn off rules that don't apply to JS code
+      '@typescript-eslint/explicit-function-return-type': 'off',
+    },
+  },
+
+  //
+  // test file linting
+  //
+
+  // test file specific configuration
+  {
+    files: ['**/*.{spec,e2e}.{ts,tsx,cts,mts}', '**/test/integration/**/*.{ts,tsx,cts,mts}'],
+    languageOptions: {
+      globals: {
+        ...jestPlugin.environments.globals.globals,
+      },
+      parserOptions: {
+        project: ['tsconfig.lint.json', 'test/**/tsconfig*.json'],
+      },
+    },
+    rules: {
+      'no-console': 'warn',
+      '@typescript-eslint/no-unsafe-argument': 'off',
+      '@typescript-eslint/await-thenable': 'off',
+      '@typescript-eslint/restrict-plus-operands': 'off',
+      '@typescript-eslint/explicit-function-return-type': 'off',
+      '@typescript-eslint/no-explicit-any': 'off',
+      '@typescript-eslint/no-empty-function': ['error', { allow: ['arrowFunctions'] }],
+      '@typescript-eslint/no-non-null-assertion': 'off',
+      '@typescript-eslint/no-unsafe-assignment': 'off',
+      '@typescript-eslint/no-unsafe-call': 'off',
+      '@typescript-eslint/no-unsafe-member-access': 'off',
+      '@typescript-eslint/no-unsafe-return': 'off',
+      'jest/no-disabled-tests': 'error',
+      'jest/no-focused-tests': 'error',
+      'jest/no-alias-methods': 'error',
+      'jest/no-identical-title': 'error',
+      'jest/no-jasmine-globals': 'error',
+      'jest/no-test-prefixes': 'error',
+      'jest/no-done-callback': 'error',
+      'jest/no-test-return-statement': 'error',
+      'jest/prefer-to-be': 'error',
+      'jest/prefer-to-contain': 'error',
+      'jest/prefer-to-have-length': 'error',
+      'jest/prefer-spy-on': 'error',
+      'jest/valid-expect': 'error',
+      'jest/no-deprecated-functions': 'error',
+    },
+  }
+);
